refactor action bar
